requirements.txt
altair==5.5.0
asn1crypto==1.5.1
attrs==25.3.0
blinker==1.9.0
boto3==1.40.24
botocore==1.40.24
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
filelock==3.19.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
llvmlite==0.44.0
MarkupSafe==3.0.2
narwhals==2.3.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.2
patsy==1.0.1
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
plotly-express==0.4.1
protobuf==6.32.0
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
s3transfer==0.13.1
scipy==1.16.2
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.17.3
sortedcontainers==2.4.0
statsmodels==0.14.5
streamlit==1.49.1
streamlit-aggrid==1.1.8.post1
streamlit-tree-select==0.0.5
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
xlsxwriter==3.2.9
